+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Add new command `init` to initialize a local workspace with a default cli.conf.yaml configuration file
+
 ## [3.1.0] - 2020-06-15
 ### Added
 - View version with `-v`